git-svn-id: file:///semplestb/final_repo@1919 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestServices/src/semplest/server/service/springjdbc/SemplestDB.java
@@ -576,7 +576,7 @@ public class SemplestDB extends BaseDB
 			//Integer FirstPageMicroCpc, Integer MicroCost
 			Integer id = setReportSP.execute(promotionID, rptObj.getKeyword(),adEngine, rptObj.getTransactionDate(), 
 					rptObj.getMicroBidAmount().intValue(),  rptObj.getNumberImpressions(), rptObj.getNumberClick(), rptObj.getAveragePosition(), rptObj.getAverageCPC(),rptObj.getBidMatchType(), 
-					rptObj.getQualityScore(), rptObj.getApprovalStatus(), rptObj.getFirstPageCPC(),
+					rptObj.getQualityScore(), rptObj.getApprovalStatus(), rptObj.getFirstPageCPC().intValue(),
 					rptObj.getMicroCost().intValue());
 			if (id == 0)
 			{
